Move view holder factory tests in the right package
