+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/views/DateTimePickerViewHolderFactoryEspressoTest.kt
@@ -33,6 +33,7 @@ import com.google.android.fhir.datacapture.TestActivity
 import com.google.android.fhir.datacapture.utilities.clickIcon
 import com.google.android.fhir.datacapture.validation.NotValidated
 import com.google.android.fhir.datacapture.views.factories.DateTimePickerViewHolderFactory
+import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolder
 import org.hamcrest.CoreMatchers.allOf
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
